projects/compliance/reporter/requirements.txt
flask
PyYAML
prometheus-client
